Separate the concepts of project and robot. (#143)
* Separate the concepts of project (which is not a module) and a robot, which is a module.
Renamed project to robot where appropriate.
Renamed module to project where appropriate.

* Make the robot's className be "Robot" everywhere, not just in the generated python code.

Added userVisibleName to the commonStorage.Project definition.

* Fixed a few errors and warnings shown in vscode.

+++ src/storage/common_storage.ts
@@ -42,10 +42,14 @@ export type Module = {
   className: string,
 };
 
+export type Robot = Module;
 export type Mechanism = Module;
 export type OpMode = Module;
 
-export type Project = Module & {
+export type Project = {
+  projectName: string, // snake_case
+  userVisibleName: string, // PascalCase
+  robot: Robot,
   mechanisms: Mechanism[]
   opModes: OpMode[],
 };
@@ -56,9 +60,11 @@ export type Component = {
 }
 
 export const MODULE_TYPE_UNKNOWN = 'unknown';
-export const MODULE_TYPE_PROJECT = 'project';
+export const MODULE_TYPE_ROBOT = 'robot';
 export const MODULE_TYPE_MECHANISM = 'mechanism';
 export const MODULE_TYPE_OPMODE = 'opmode';
+
+export const ROBOT_CLASS_NAME = 'Robot';
 
 export const MODULE_NAME_PLACEHOLDER = '%module_name%';
 
@@ -78,15 +84,69 @@ export const UPLOAD_DOWNLOAD_FILE_EXTENSION = '.blocks';
 export interface Storage {
   saveEntry(entryKey: string, entryValue: string): Promise<void>;
   fetchEntry(entryKey: string, defaultValue: string): Promise<string>;
-  listModules(): Promise<Project[]>;
+  listProjects(): Promise<Project[]>;
   fetchModuleContent(modulePath: string): Promise<string>;
+  createProject(projectName: string, robotContent: string): Promise<void>;
   createModule(moduleType: string, modulePath: string, moduleContent: string): Promise<void>;
   saveModule(modulePath: string, moduleContent: string): Promise<void>;
+  renameProject(oldProjectName: string, newProjectName: string): Promise<void>;
+  copyProject(oldProjectName: string, newProjectName: string): Promise<void>;
   renameModule(moduleType: string, projectName: string, oldModuleName: string, newModuleName: string): Promise<void>;
   copyModule(moduleType: string, projectName: string, oldModuleName: string, newModuleName: string): Promise<void>;
+  deleteProject(projectName: string): Promise<void>;
   deleteModule(moduleType: string, modulePath: string): Promise<void>;
   downloadProject(projectName: string): Promise<string>;
   uploadProject(projectName: string, blobUrl: string): Promise<void>;
+}
+
+/**
+ * Creates a new project.
+ * @param storage The storage interface to use for creating the project.
+ * @param proposedUserVisibleName The name for the new project.
+ * @returns A promise that resolves when the project has been created.
+ */
+export async function createProject(
+  storage: Storage, proposedUserVisibleName: string): Promise<void> {
+  const newProjectName = pascalCaseToSnakeCase(proposedUserVisibleName);
+  const robotContent = newRobotContent(newProjectName);
+  await storage.createProject(newProjectName, robotContent);
+}
+
+/**
+ * Renames a project.
+ * @param storage The storage interface to use for renaming the project.
+ * @param project The project to rename
+ * @param proposedUserVisibleName The new name for the project.
+ * @returns A promise that resolves when the project has been renamed.
+ */
+export async function renameProject(
+  storage: Storage, project: Project, proposedUserVisibleName: string): Promise<void> {
+  const newProjectName = pascalCaseToSnakeCase(proposedUserVisibleName);
+  await storage.renameProject(project.projectName, newProjectName);
+}
+
+/**
+ * Copies a project.
+ * @param storage The storage interface to use for copying the project.
+ * @param project The project to copy
+ * @param proposedUserVisibleName The name for the new project.
+ * @returns A promise that resolves when the project has been copied.
+ */
+export async function copyProject(
+  storage: Storage, project: Project, proposedUserVisibleName: string): Promise<void> {
+  const newProjectName = pascalCaseToSnakeCase(proposedUserVisibleName);
+  await storage.copyProject(project.projectName, newProjectName);
+}
+
+/**
+ * Deletes a project.
+ * @param storage The storage interface to use for deleting the project.
+ * @param project The project to delete.
+ * @returns A promise that resolves when the project has been deleted.
+ */
+export async function deleteProject(
+  storage: Storage, project: Project): Promise<void> {
+  await storage.deleteProject(project.projectName);
 }
 
 /**
@@ -94,11 +154,11 @@ export interface Storage {
  * @param storage The storage interface to use for creating the module.
  * @param project The project to add the module to.
  * @param moduleType The type of the module (e.g., 'mechanism', 'opmode').
- * @param className The name of the class.
+ * @param newClassName The name of the class.
  */
 export async function addModuleToProject(
-  storage: Storage, project: Project, moduleType: string, className: string): Promise<void> {
-  let newModuleName = classNameToModuleName(className);
+  storage: Storage, project: Project, moduleType: string, newClassName: string): Promise<void> {
+  let newModuleName = pascalCaseToSnakeCase(newClassName);
   let newModulePath = makeModulePath(project.projectName, newModuleName);
 
   if (moduleType === MODULE_TYPE_MECHANISM) {
@@ -109,7 +169,7 @@ export async function addModuleToProject(
       moduleType: MODULE_TYPE_MECHANISM,
       projectName: project.projectName,
       moduleName: newModuleName,
-      className: className
+      className: newClassName
     } as Mechanism);
   } else if (moduleType === MODULE_TYPE_OPMODE) {
     const opModeContent = newOpModeContent(project.projectName, newModuleName);
@@ -119,7 +179,7 @@ export async function addModuleToProject(
       moduleType: MODULE_TYPE_OPMODE,
       projectName: project.projectName,
       moduleName: newModuleName,
-      className: className
+      className: newClassName
     } as OpMode);
   }
 }
@@ -131,8 +191,11 @@ export async function addModuleToProject(
  */
 export async function removeModuleFromProject(
   storage: Storage, project: Project, modulePath: string): Promise<void> {
-  const module = findModuleInProject(project, modulePath);
+  const module = findModuleByModulePath(project, modulePath);
   if (module) {
+    if (module.moduleType == MODULE_TYPE_ROBOT) {
+      throw new Error('Removing the robot module from the project is not allowed.');
+    }
     await storage.deleteModule(module.moduleType, modulePath);
     if (module.moduleType === MODULE_TYPE_MECHANISM) {
       project.mechanisms = project.mechanisms.filter(m => m.modulePath !== modulePath);
@@ -146,34 +209,37 @@ export async function removeModuleFromProject(
  * Renames a module in the project.
  * @param storage The storage interface to use for renaming the module.
  * @param project The project containing the module to rename.
- * @param proposedName The new name for the module.
- * @param oldModuleName The current name of the module.
+ * @param proposedClassName The new class name for the module.
+ * @param oldModulePath The current path of the module.
  * @returns A promise that resolves when the module has been renamed.
  */
 export async function renameModuleInProject(
-  storage: Storage, project: Project, proposedName: string, oldModulePath: string): Promise<string> {
-  const module = findModuleInProject(project, oldModulePath);
+  storage: Storage, project: Project, proposedClassName: string, oldModulePath: string): Promise<string> {
+  const module = findModuleByModulePath(project, oldModulePath);
   if (module) {
-    const newModuleName = classNameToModuleName(proposedName);
+    if (module.moduleType == MODULE_TYPE_ROBOT) {
+      throw new Error('Renaming the robot module is not allowed.');
+    }
+    const newModuleName = pascalCaseToSnakeCase(proposedClassName);
     const newModulePath = makeModulePath(project.projectName, newModuleName);
     await storage.renameModule(module.moduleType, project.projectName, module.moduleName, newModuleName);
     module.modulePath = newModulePath;
     module.moduleName = newModuleName;
-    module.className = proposedName;
+    module.className = proposedClassName;
 
     if (module.moduleType === MODULE_TYPE_MECHANISM) {
       const mechanism = project.mechanisms.find(m => m.modulePath === module.modulePath);
       if (mechanism) {
         mechanism.modulePath = newModulePath;
         mechanism.moduleName = newModuleName;
-        mechanism.className = proposedName;
+        mechanism.className = proposedClassName;
       }
     } else if (module.moduleType === MODULE_TYPE_OPMODE) {
       const opMode = project.opModes.find(o => o.modulePath === module.modulePath);
       if (opMode) {
         opMode.modulePath = newModulePath;
         opMode.moduleName = newModuleName;
-        opMode.className = proposedName;
+        opMode.className = proposedClassName;
       }
       return newModulePath
     }
@@ -184,15 +250,18 @@ export async function renameModuleInProject(
  * Copies a module in the project.
  * @param storage The storage interface to use for copying the module.
  * @param project The project containing the module to copy.
- * @param proposedName The new name for the module.
+ * @param proposedClassName The new name for the module.
  * @param oldModuleName The current name of the module.
  * @returns A promise that resolves when the module has been copied.
  */
 export async function copyModuleInProject(
-  storage: Storage, project: Project, proposedName: string, oldModulePath: string): Promise<string> {
-  const module = findModuleInProject(project, oldModulePath);
+  storage: Storage, project: Project, proposedClassName: string, oldModulePath: string): Promise<string> {
+  const module = findModuleByModulePath(project, oldModulePath);
   if (module) {
-    const newModuleName = classNameToModuleName(proposedName);
+    if (module.moduleType == MODULE_TYPE_ROBOT) {
+      throw new Error('Copying the robot module is not allowed.');
+    }
+    const newModuleName = pascalCaseToSnakeCase(proposedClassName);
     const newModulePath = makeModulePath(project.projectName, newModuleName);
     await storage.copyModule(module.moduleType, project.projectName, module.moduleName, newModuleName);
 
@@ -202,7 +271,7 @@ export async function copyModuleInProject(
         moduleType: MODULE_TYPE_MECHANISM,
         projectName: project.projectName,
         moduleName: newModuleName,
-        className: proposedName
+        className: proposedClassName
       } as Mechanism);
     } else if (module.moduleType === MODULE_TYPE_OPMODE) {
       project.opModes.push({
@@ -210,7 +279,7 @@ export async function copyModuleInProject(
         moduleType: MODULE_TYPE_OPMODE,
         projectName: project.projectName,
         moduleName: newModuleName,
-        className: proposedName
+        className: proposedClassName
       } as OpMode);
     }
     return newModulePath;
@@ -221,24 +290,19 @@ export async function copyModuleInProject(
 /**
  * Checks if the proposed class name is valid and does not conflict with existing names in the project.
  * @param project The project to check against.
- * @param proposedName The proposed class name to validate.
+ * @param proposedClassName The proposed class name to validate.
  * @returns An object containing a boolean `ok` indicating if the name is valid, and an `error` message if it is not.
  */
-export function isClassNameOk(project: Project, proposedName: string) {
+export function isClassNameOk(project: Project, proposedClassName: string) {
   let ok = true;
   let error = '';
 
-  if (!isValidClassName(proposedName)) {
+  if (!isValidClassName(proposedClassName)) {
     ok = false;
-    error = proposedName + ' is not a valid name. Please enter a different name.';
-  }
-  else if (proposedName == project.className) {
+    error = proposedClassName + ' is not a valid name. Please enter a different name.';
+  } else if (findModuleByClassName(project, proposedClassName) != null) {
     ok = false;
-    error = 'The project is already named ' + proposedName + '. Please enter a different name.';
-  }
-  else if (getClassInProject(project, proposedName) != null) {
-    ok = false;
-    error = 'Another Mechanism or OpMode is already named ' + proposedName + '. Please enter a different name.'
+    error = 'Another Mechanism or OpMode is already named ' + proposedClassName + '. Please enter a different name.'
   }
 
   return {
@@ -248,16 +312,19 @@ export function isClassNameOk(project: Project, proposedName: string) {
 }
 
 /**
- * Returns true if the given classname is in the project
+ * Returns the module in the given project with the given class name.
  */
-export function getClassInProject(project: Project, name: string): Module | null {
+export function findModuleByClassName(project: Project, className: string): Module | null {
+  if (project.robot.className === className) {
+    return project.robot;
+  }
   for (const mechanism of project.mechanisms) {
-    if (mechanism.className === name) {
+    if (mechanism.className === className) {
       return mechanism;
     }
   }
   for (const opMode of project.opModes) {
-    if (opMode.className === name) {
+    if (opMode.className === className) {
       return opMode;
     }
   }
@@ -265,25 +332,11 @@ export function getClassInProject(project: Project, name: string): Module | null
 }
 
 /**
- * Returns the module with the given module path, or null if it is not found.
- */
-export function findModule(modules: Project[], modulePath: string): Module | null {
-  for (const project of modules) {
-    const result = findModuleInProject(project, modulePath);
-    if (result) {
-      return result;
-    }
-  }
-
-  return null;
-}
-
-/**
  * Returns the module with the given module path inside the given project, or null if it is not found.
  */
-export function findModuleInProject(project: Project, modulePath: string): Module | null {
-  if (project.modulePath === modulePath) {
-    return project;
+export function findModuleByModulePath(project: Project, modulePath: string): Module | null {
+  if (project.robot.modulePath === modulePath) {
+    return project.robot;
   }
   for (const mechanism of project.mechanisms) {
     if (mechanism.modulePath === modulePath) {
@@ -343,9 +396,9 @@ export function isValidClassName(name: string): boolean {
 }
 
 /**
- * Returns the module name for the given class name.
+ * Returns the module name (snake_case) for the given class name (PascalCase).
  */
-export function classNameToModuleName(className: string): string {
+export function pascalCaseToSnakeCase(className: string): string {
   let moduleName = '';
   for (let i = 0; i < className.length; i++) {
     const char = className.charAt(i);
@@ -362,9 +415,9 @@ export function classNameToModuleName(className: string): string {
 }
 
 /**
- * Returns the class name for the given module name.
+ * Returns the class name (PascalCase) for the given module name (snake_case).
  */
-export function moduleNameToClassName(moduleName: string): string {
+export function snakeCaseToPascalCase(moduleName: string): string {
   let className = '';
   let nextCharUpper = true;
   for (let i = 0; i < moduleName.length; i++) {
@@ -395,14 +448,14 @@ export function makeModulePath(projectName: string, moduleName: string): string 
 }
 
 /**
- * Returns the project path for the given project names.
+ * Returns the robot module path for the given project names.
  */
-export function makeProjectPath(projectName: string): string {
+export function makeRobotPath(projectName: string): string {
   return makeModulePath(projectName, projectName);
 }
 
 /**
- * Returns the project name for given module path.
+ * Returns the project path for given module path.
  */
 export function getProjectName(modulePath: string): string {
   const regex = new RegExp('^([a-z_A-Z][a-z0-9_]*)/([a-z_A-Z][a-z0-9_]*).py$');
@@ -446,16 +499,16 @@ function startingBlocksToModuleContent(
 }
 
 /**
- * Returns the module content for a new Project.
+ * Returns the robot module content for a new Project.
  */
-export function newProjectContent(projectName: string): string {
-  const module: Module = {
-    modulePath: makeProjectPath(projectName),
-    moduleType: MODULE_TYPE_PROJECT,
+export function newRobotContent(projectName: string): string {
+  const module: Robot = {
+    modulePath: makeRobotPath(projectName),
+    moduleType: MODULE_TYPE_ROBOT,
     projectName: projectName,
     moduleName: projectName,
     dateModifiedMillis: 0,
-    className: moduleNameToClassName(projectName),
+    className: ROBOT_CLASS_NAME,
   };
 
   return startingBlocksToModuleContent(module, startingRobotBlocks);
@@ -471,7 +524,7 @@ export function newMechanismContent(projectName: string, mechanismName: string):
     projectName: projectName,
     moduleName: mechanismName,
     dateModifiedMillis: 0,
-    className: moduleNameToClassName(mechanismName),
+    className: snakeCaseToPascalCase(mechanismName),
   };
 
   return startingBlocksToModuleContent(module, startingMechanismBlocks);
@@ -487,7 +540,7 @@ export function newOpModeContent(projectName: string, opModeName: string): strin
     projectName: projectName,
     moduleName: opModeName,
     dateModifiedMillis: 0,
-    className: moduleNameToClassName(opModeName),
+    className: snakeCaseToPascalCase(opModeName),
   };
 
   return startingBlocksToModuleContent(module, startingOpModeBlocks);
@@ -640,6 +693,12 @@ export async function produceDownloadProjectBlob(
   return blobUrl;
 }
 
+export function getClassNameForModule(moduleType: string, moduleName: string) {
+  return (moduleType == MODULE_TYPE_ROBOT)
+      ? ROBOT_CLASS_NAME
+      : snakeCaseToPascalCase(moduleName);
+}
+
 /**
  * Process the module content so it can be downloaded.
  */
@@ -661,7 +720,7 @@ function _processModuleContentForDownload(
     projectName: projectName,
     moduleName: moduleName,
     dateModifiedMillis: 0,
-    className: moduleNameToClassName(moduleName),
+    className: getClassNameForModule(moduleType, moduleName),
   };
 
   // Clear out the python content and exported blocks.
@@ -754,7 +813,7 @@ export function _processUploadedModule(
     moduleType = moduleType.substring(MARKER_MODULE_TYPE.length);
   }
 
-  const moduleName = (moduleType === MODULE_TYPE_PROJECT)
+  const moduleName = (moduleType === MODULE_TYPE_ROBOT)
     ? projectName : filename;
 
   const module: Module = {
@@ -763,7 +822,7 @@ export function _processUploadedModule(
     projectName: projectName,
     moduleName: moduleName,
     dateModifiedMillis: 0,
-    className: moduleNameToClassName(moduleName),
+    className: snakeCaseToPascalCase(moduleName),
   };
 
   // Generate the python content and exported blocks.
